move: atomsにボタンディレクトリを作成
