+++ src/components/pages/auth/Login.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, FC, memo, useCallback, useState } from "react";
 import { Box, Flex, Heading, Link } from "@chakra-ui/react";
 
-import { PrimaryButton } from "../../atoms/PrimaryButton";
+import { PrimaryButton } from "../../atoms/buttons/PrimaryButton";
 import { PasswordInput } from "../../molecules/PasswordInput";
 import { useSignIn } from "../../../hooks/useSignIn";
 import { PrimaryInputArea } from "../../molecules/PrimaryInputArea";
