+++ src/components/pages/auth/SignUp.tsx
@@ -3,7 +3,7 @@ import { Box, Flex, Heading, Text } from "@chakra-ui/layout";
 import { Avatar } from "@chakra-ui/avatar";
 
 import { PasswordInput } from "../../molecules/PasswordInput";
-import { PrimaryButton } from "../../atoms/PrimaryButton";
+import { PrimaryButton } from "../../atoms/buttons/PrimaryButton";
 import { PrimaryInputArea } from "../../molecules/PrimaryInputArea";
 import { UploadPhotoButton } from "../../molecules/UploadPhotoButton";
 import { useSignUp } from "../../../hooks/useSignUp";
